+++ packages/core/base/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @solana/wallet-adapter-base
 
+## 0.9.22
+
+### Patch Changes
+
+-   3f73508d: Update dependencies
+
+## 0.9.21
+
+### Patch Changes
+
+-   f99c2154: Add StandardAdapter to base Adapter type
+
 ## 0.9.20
 
 ### Patch Changes