+++ packages/wallets/ledger/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @solana/wallet-adapter-ledger
 
+## 0.9.24
+
+### Patch Changes
+
+-   3f73508d: Update dependencies
+-   Updated dependencies [3f73508d]
+    -   @solana/wallet-adapter-base@0.9.22
+
+## 0.9.23
+
+### Patch Changes
+
+-   f99c2154: Add v0 tx support to Ledger adapter
+-   Updated dependencies [f99c2154]
+    -   @solana/wallet-adapter-base@0.9.21
+
 ## 0.9.22
 
 ### Patch Changes