+++ packages/wallets/solflare/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @solana/wallet-adapter-solflare
 
+## 0.6.24
+
+### Patch Changes
+
+-   3f73508d: Update dependencies
+-   Updated dependencies [3f73508d]
+    -   @solana/wallet-adapter-base@0.9.22
+
+## 0.6.23
+
+### Patch Changes
+
+-   Updated dependencies [f99c2154]
+    -   @solana/wallet-adapter-base@0.9.21
+
 ## 0.6.22
 
 ### Patch Changes