+++ packages/wallets/sollet/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @solana/wallet-adapter-sollet
 
+## 0.11.16
+
+### Patch Changes
+
+-   3f73508d: Update dependencies
+-   Updated dependencies [3f73508d]
+    -   @solana/wallet-adapter-base@0.9.22
+
+## 0.11.15
+
+### Patch Changes
+
+-   Updated dependencies [f99c2154]
+    -   @solana/wallet-adapter-base@0.9.21
+
 ## 0.11.14
 
 ### Patch Changes