+++ packages/wallets/walletconnect/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @solana/wallet-adapter-walletconnect
 
+## 0.1.14
+
+### Patch Changes
+
+-   3f73508d: Update dependencies
+-   Updated dependencies [3f73508d]
+    -   @solana/wallet-adapter-base@0.9.22
+
+## 0.1.13
+
+### Patch Changes
+
+-   61e86b58: Taking a new version of `@jnwng/walletconnect-solana` with upgraded dependencies on `@walletconnect/sign-client`
+
+## 0.1.12
+
+### Patch Changes
+
+-   f5abf15c: Using WalletConnect (alone, or through `@solana/wallet-adapter-wallets` no longer fatals in Next 13.
+
+## 0.1.11
+
+### Patch Changes
+
+-   Updated dependencies [f99c2154]
+    -   @solana/wallet-adapter-base@0.9.21
+
 ## 0.1.10
 
 ### Patch Changes